+++ src/DependencyInjection/Configuration.php
@@ -10,16 +10,17 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 class Configuration implements ConfigurationInterface
 {
-    public const NAME_ATTRIBUTE = 'name';
+    public const USER_ATTRIBUTE_NAME_ATTRIBUTE = 'name';
 
     public const USER_ATTRIBUTE_PROVIDER_ATTRIBUTE = 'user_attribute_provider';
-    public const ATTRIBUTES_ATTRIBUTE = 'attributes';
-    public const LDAP_ATTRIBUTE_ATTRIBUTE = 'ldap_attribute';
-    public const IS_ARRAY_ATTRIBUTE = 'is_array';
-    public const DEFAULT_VALUE_ATTRIBUTE = 'default_value';
-    public const DEFAULT_VALUES_ATTRIBUTE = 'default_values';
-    public const LDAP_CONNECTION_ATTRIBUTE = 'ldap_connection';
-    public const LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE = 'ldap_user_identifier_attribute';
+    public const USER_ATTRIBUTE_ATTRIBUTES_ATTRIBUTE = 'attributes';
+    public const USER_ATTRIBUTE_LDAP_CONNECTION_ATTRIBUTE = 'ldap_connection';
+    public const USER_ATTRIBUTE_LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE = 'ldap_user_identifier_attribute';
+    public const USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_ATTRIBUTE = 'current_ldap_user_identifier_expression';
+    public const USER_ATTRIBUTE_LDAP_ATTRIBUTE_ATTRIBUTE = 'ldap_attribute';
+    public const USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE = 'is_array';
+    public const USER_ATTRIBUTE_DEFAULT_VALUE_ATTRIBUTE = 'default_value';
+    public const USER_ATTRIBUTE_DEFAULT_VALUES_ATTRIBUTE = 'default_values';
 
     public const CONNECTIONS_ATTRIBUTE = 'connections';
     public const LDAP_CONNECTION_IDENTIFIER_ATTRIBUTE = 'identifier';
@@ -34,6 +35,8 @@ class Configuration implements ConfigurationInterface
 
     /** @var int */
     public const LDAP_NUM_RESULT_ITEMS_WILL_SORT_LIMIT_DEFAULT = 10000;
+
+    private const USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_DEFAULT_VALUE = 'user.getIdentifier()';
 
     public function getConfigTreeBuilder(): TreeBuilder
     {
@@ -55,39 +58,45 @@ class Configuration implements ConfigurationInterface
 
         return $treeBuilder->getRootNode()
             ->children()
-                ->arrayNode(self::ATTRIBUTES_ATTRIBUTE)
+                ->arrayNode(self::USER_ATTRIBUTE_ATTRIBUTES_ATTRIBUTE)
                     ->arrayPrototype()
                         ->children()
-                            ->scalarNode(self::NAME_ATTRIBUTE)
+                            ->scalarNode(self::USER_ATTRIBUTE_NAME_ATTRIBUTE)
                                 ->cannotBeEmpty()
                                 ->isRequired()
                                 ->info('The name of the authorization attribute')
                             ->end()
-                            ->scalarNode(self::LDAP_ATTRIBUTE_ATTRIBUTE)
+                            ->scalarNode(self::USER_ATTRIBUTE_LDAP_ATTRIBUTE_ATTRIBUTE)
                                 ->cannotBeEmpty()
                                 ->isRequired()
                                 ->info('The source LDAP attribute that is mapped to the authorization attribute. If left blank, the attribute\'s value is not automatically mapped.')
                             ->end()
-                            ->booleanNode(self::IS_ARRAY_ATTRIBUTE)
+                            ->booleanNode(self::USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE)
                                 ->defaultFalse()
                             ->end()
-                            ->scalarNode(self::DEFAULT_VALUE_ATTRIBUTE)
+                            ->scalarNode(self::USER_ATTRIBUTE_DEFAULT_VALUE_ATTRIBUTE)
                                 ->info('The default value for scalar (non-array) attributes. The default is null.')
                             ->end()
-                            ->arrayNode(self::DEFAULT_VALUES_ATTRIBUTE)
+                            ->arrayNode(self::USER_ATTRIBUTE_DEFAULT_VALUES_ATTRIBUTE)
                                 ->info('The default value for array type attributes. The default is an empty array.')
                                 ->scalarPrototype()->end()
                             ->end()
                         ->end()
                     ->end()
                 ->end()
-                ->scalarNode(self::LDAP_CONNECTION_ATTRIBUTE)
+                ->scalarNode(self::USER_ATTRIBUTE_LDAP_CONNECTION_ATTRIBUTE)
                    ->cannotBeEmpty()
                    ->isRequired()
                    ->info('The identifier of the LDAP connection to use. See the dbp_relay_ldap config for available connections.')
                 ->end()
-                ->scalarNode(self::LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE)
+                ->scalarNode(self::USER_ATTRIBUTE_LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE)
                    ->defaultValue('cn')
+                   ->info('The LDAP attribute to use for the lookup of the current user identifier. The default is "cn".')
+                ->end()
+                ->scalarNode(self::USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_ATTRIBUTE)
+                    ->defaultValue(self::USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_DEFAULT_VALUE)
+                    ->info('The expression that is evaluated to get the current user identifier. The default is '.
+                        self::USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_DEFAULT_VALUE)
                 ->end()
             ->end()
         ;