+++ tests/UserAttributeProviderTest.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\CoreConnectorLdapBundle\Tests;
 
+use Dbp\Relay\CoreBundle\TestUtils\TestAuthorizationService;
 use Dbp\Relay\CoreBundle\TestUtils\TestUserSession;
 use Dbp\Relay\CoreBundle\User\UserAttributeException;
 use Dbp\Relay\CoreConnectorLdapBundle\DependencyInjection\Configuration;
+use Dbp\Relay\CoreConnectorLdapBundle\Service\AuthorizationService;
 use Dbp\Relay\CoreConnectorLdapBundle\Service\UserAttributeProvider;
 use Dbp\Relay\CoreConnectorLdapBundle\TestUtils\TestLdapConnectionProvider;
 use PHPUnit\Framework\TestCase;
@@ -21,8 +23,8 @@ class UserAttributeProviderTest extends TestCase
     public const LDAP_IDENTIFIER_ATTRIBUTE_NAME = 'cn';
     public const LDAP_ROLES_ATTRIBUTE_NAME = 'LDAP-ROLES';
 
-    private ?UserAttributeProvider $userAttributeProvider;
-    private ?TestLdapConnectionProvider $testLdapConnectionProvider;
+    private ?UserAttributeProvider $userAttributeProvider = null;
+    private ?TestLdapConnectionProvider $testLdapConnectionProvider = null;
 
     protected function setUp(): void
     {
@@ -31,39 +33,46 @@ class UserAttributeProviderTest extends TestCase
 
     protected function tearDown(): void
     {
-        $this->testLdapConnectionProvider->cleanup();
+        $this->testLdapConnectionProvider->tearDown();
     }
 
-    private function setupUserAttributeProviderWithUser(string $userIdentifier = 'testuser', bool $isAuthenticated = true, bool $isServiceAccount = false): void
+    private function setupUserAttributeProviderWithUser(?string $userIdentifier = 'testuser'): void
     {
+        $config = $this->createConfig();
+        $authorizationService = new AuthorizationService();
+        TestAuthorizationService::setUp($authorizationService, $userIdentifier);
+        $authorizationService->setConfig($config[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE]);
+
         $this->userAttributeProvider = new UserAttributeProvider(
-            $this->testLdapConnectionProvider, new TestUserSession(
-                $userIdentifier, isAuthenticated: $isAuthenticated, isServiceAccount: $isServiceAccount));
-        $this->userAttributeProvider->setConfig($this->createConfig()[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE]);
+            $this->testLdapConnectionProvider,
+            new TestUserSession($userIdentifier, isAuthenticated: true),
+            $authorizationService);
+        $this->userAttributeProvider->setConfig($config[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE]);
     }
 
     private function createConfig(): array
     {
         return [
             Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE => [
-                Configuration::LDAP_CONNECTION_ATTRIBUTE => TestLdapConnectionProvider::DEFAULT_CONNECTION_IDENTIFIER,
-                Configuration::LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE => self::LDAP_IDENTIFIER_ATTRIBUTE_NAME,
-                Configuration::ATTRIBUTES_ATTRIBUTE => [
+                Configuration::USER_ATTRIBUTE_LDAP_CONNECTION_ATTRIBUTE => TestLdapConnectionProvider::DEFAULT_CONNECTION_IDENTIFIER,
+                Configuration::USER_ATTRIBUTE_LDAP_USER_IDENTIFIER_ATTRIBUTE_ATTRIBUTE => self::LDAP_IDENTIFIER_ATTRIBUTE_NAME,
+                Configuration::USER_ATTRIBUTE_CURRENT_LDAP_USER_IDENTIFIER_EXPRESSION_ATTRIBUTE => 'user.getIdentifier()',
+                Configuration::USER_ATTRIBUTE_ATTRIBUTES_ATTRIBUTE => [
                     [
-                        Configuration::NAME_ATTRIBUTE => self::ROLES_ATTRIBUTE,
-                        Configuration::IS_ARRAY_ATTRIBUTE => true,
-                        Configuration::LDAP_ATTRIBUTE_ATTRIBUTE => self::LDAP_ROLES_ATTRIBUTE_NAME,
-                        Configuration::DEFAULT_VALUES_ATTRIBUTE => ['DEFAULT'],
+                        Configuration::USER_ATTRIBUTE_NAME_ATTRIBUTE => self::ROLES_ATTRIBUTE,
+                        Configuration::USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE => true,
+                        Configuration::USER_ATTRIBUTE_LDAP_ATTRIBUTE_ATTRIBUTE => self::LDAP_ROLES_ATTRIBUTE_NAME,
+                        Configuration::USER_ATTRIBUTE_DEFAULT_VALUES_ATTRIBUTE => ['DEFAULT'],
                     ],
                     [
-                        Configuration::NAME_ATTRIBUTE => self::MISC_ATTRIBUTE,
-                        Configuration::IS_ARRAY_ATTRIBUTE => false,
-                        Configuration::DEFAULT_VALUE_ATTRIBUTE => 0,
+                        Configuration::USER_ATTRIBUTE_NAME_ATTRIBUTE => self::MISC_ATTRIBUTE,
+                        Configuration::USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE => false,
+                        Configuration::USER_ATTRIBUTE_DEFAULT_VALUE_ATTRIBUTE => 0,
                     ],
                     [
-                        Configuration::NAME_ATTRIBUTE => self::MISC_ARRAY_ATTRIBUTE,
-                        Configuration::IS_ARRAY_ATTRIBUTE => true,
-                        Configuration::DEFAULT_VALUES_ATTRIBUTE => [1, 2, 3],
+                        Configuration::USER_ATTRIBUTE_NAME_ATTRIBUTE => self::MISC_ARRAY_ATTRIBUTE,
+                        Configuration::USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE => true,
+                        Configuration::USER_ATTRIBUTE_DEFAULT_VALUES_ATTRIBUTE => [1, 2, 3],
                     ],
                 ],
             ],
@@ -99,20 +108,17 @@ class UserAttributeProviderTest extends TestCase
         $this->mockResultsFor('money82');
         $authzUserAttributes = $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE);
 
-        $this->mockResultsFor('foo');
+        $this->mockResultsFor('money82');
         $authzUserAttributeCached = $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE);
-        $this->assertNotSame($authzUserAttributes, $authzUserAttributeCached);
+        $this->assertSame($authzUserAttributes, $authzUserAttributeCached);
 
         // with caching
         $this->userAttributeProvider->setCache(new ArrayAdapter());
-        $this->mockResultsFor('money82');
-        $authzUserAttribute = $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE);
-        $this->assertEquals(['VIEWER', 'EDITOR'], $authzUserAttribute);
 
-        $this->mockResultsFor('foo');
+        $this->mockResultsFor('money82');
+        $authzUserAttributes = $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE);
         // no new LDAP request since user is found in cache
         $authzUserAttributeCached = $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE);
-
         $this->assertSame($authzUserAttributes, $authzUserAttributeCached);
     }
 
@@ -144,34 +150,11 @@ class UserAttributeProviderTest extends TestCase
     public function testWithoutUserId()
     {
         // expecting all default values
-        $this->setupUserAttributeProviderWithUser();
+        $this->setupUserAttributeProviderWithUser(null);
 
         $this->assertEquals(['DEFAULT'], $this->userAttributeProvider->getUserAttribute(null, self::ROLES_ATTRIBUTE));
         $this->assertEquals(0, $this->userAttributeProvider->getUserAttribute(null, self::MISC_ATTRIBUTE));
         $this->assertEquals([1, 2, 3], $this->userAttributeProvider->getUserAttribute(null, self::MISC_ARRAY_ATTRIBUTE));
-    }
-
-    /**
-     * @throws UserAttributeException
-     */
-    public function testUnauthenticated()
-    {
-        // this is allowed for debugging purposes
-        $this->setupUserAttributeProviderWithUser(isAuthenticated: false);
-        $this->mockResultsFor('money82');
-        $this->assertEquals(['VIEWER', 'EDITOR'],
-            $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE));
-    }
-
-    /**
-     * @throws UserAttributeException
-     */
-    public function testUserIdentifierMismatch()
-    {
-        // it's allowed to request attributes for users other than the logged-in user
-        $this->setupUserAttributeProviderWithUser('foo');
-        $this->mockResultsFor('money82');
-        $this->assertEquals(['VIEWER', 'EDITOR'], $this->userAttributeProvider->getUserAttribute('money82', self::ROLES_ATTRIBUTE));
     }
 
     public function testUserNotFound()
@@ -180,7 +163,11 @@ class UserAttributeProviderTest extends TestCase
         $this->mockResultsFor();
         // expecting all default values
         $this->assertEquals(['DEFAULT'], $this->userAttributeProvider->getUserAttribute('foo', self::ROLES_ATTRIBUTE));
+
+        $this->mockResultsFor();
         $this->assertEquals(0, $this->userAttributeProvider->getUserAttribute('foo', self::MISC_ATTRIBUTE));
+
+        $this->mockResultsFor();
         $this->assertEquals([1, 2, 3], $this->userAttributeProvider->getUserAttribute('foo', self::MISC_ARRAY_ATTRIBUTE));
     }
 
@@ -190,10 +177,10 @@ class UserAttributeProviderTest extends TestCase
 
         $config = $this->createConfig();
         // add duplicate attribute entry
-        $config[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE][Configuration::ATTRIBUTES_ATTRIBUTE][] = [
-            Configuration::NAME_ATTRIBUTE => self::ROLES_ATTRIBUTE,
-            Configuration::IS_ARRAY_ATTRIBUTE => true,
-            Configuration::LDAP_ATTRIBUTE_ATTRIBUTE => self::LDAP_ROLES_ATTRIBUTE_NAME,
+        $config[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE][Configuration::USER_ATTRIBUTE_ATTRIBUTES_ATTRIBUTE][] = [
+            Configuration::USER_ATTRIBUTE_NAME_ATTRIBUTE => self::ROLES_ATTRIBUTE,
+            Configuration::USER_ATTRIBUTE_IS_ARRAY_ATTRIBUTE => true,
+            Configuration::USER_ATTRIBUTE_LDAP_ATTRIBUTE_ATTRIBUTE => self::LDAP_ROLES_ATTRIBUTE_NAME,
         ];
         $this->expectException(\RuntimeException::class);
         $this->userAttributeProvider->setConfig($config[Configuration::USER_ATTRIBUTE_PROVIDER_ATTRIBUTE]);
@@ -231,6 +218,6 @@ class UserAttributeProviderTest extends TestCase
             default => [],
         };
 
-        $this->testLdapConnectionProvider->mockResults($results);
+        $this->testLdapConnectionProvider->mockResults($results, $userIdentifier);
     }
 }